+++ src/StratDiff.hpp
@@ -57,5 +57,5 @@ public:
     \param allCards The vector of all our cards, including the dog.
     \return A set of Card pointers for the cards we place into the ecart.
   */
-  virtual set< shared_ptr<Card> > makeEcart( const int dogSize, const vector< shared_ptr<Card> >& allCards ) const = 0;
+  virtual set< shared_ptr<Card> > makeEcart( const size_t dogSize, const vector< shared_ptr<Card> >& allCards ) const = 0;
 };
